cve_project/requirements.txt
arabic-reshaper==3.0.0
arpy==2.3.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
cssselect2==0.7.0
cx_Freeze==7.2.10
dateutils==0.6.12
Django==5.1.6
django-environ==0.12.0
django-weasyprint==2.3.1
Events==0.5
filelock==3.17.0
fonttools==4.56.0
gunicorn==23.0.0
html5lib==1.1
idna==3.10
lxml==5.3.0
opensearch-dsl==2.1.0
opensearch-py==2.8.0
oscrypto==1.3.0
packaging==24.2
patchelf==0.17.2.1
PGPy==0.6.0
pillow==11.1.0
psycopg==3.2.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydpkg==1.9.3
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pypdf==5.2.0
pyphen==0.17.2
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-debian==0.1.49
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
reportlab==4.3.0
requests==2.32.3
rpm_vercmp==0.1.2
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.3.0
weasyprint==64.0
webencodings==0.5.1
xmltodict==0.14.2
zopfli==0.2.3.post1
zstandard==0.22.0
